requirements.txt
PySide6>=6.4.0
Pillow>=9.3.0
mss>=6.1.0
google-generativeai>=0.3.0
pynput>=1.7.0
python-dotenv>=0.21.0
pyobjc-framework-Quartz>=9.0  # For macOS screen capture
pyobjc-framework-ApplicationServices>=9.0  # For macOS UI automation